pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "fake-api-server"
version = "0.2.0"
description = "🕸🤖👺 A Python tool to fake API server, e.g., RESTful API server, easily and humanly without any coding."
authors = [
    { name = "Liu, Bryant", email = "chi10211201@cycu.org.tw" }
]
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["mock", "API", "mock server", "mock API server", "web", "server"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Internet",
    "Topic :: Utilities",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing :: Mocking",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8,<3.13"
dependencies = [
    "PyYAML (>=6.0)",
]

[tool.poetry]
name = "fake-api-server"
version = "0.2.0"
description = "🕸🤖👺 A Python tool to fake API server, e.g., RESTful API server, easily and humanly without any coding."
authors = [
    "Liu, Bryant <chi10211201@cycu.org.tw>",
]
packages = [
    { include = "fake_api_server/" },
]

[project.optional-dependencies]
mini = [
    "PyYAML (>=6.0)",
]
fastapi = [
    "PyYAML (>=6.0)",
    "fastapi (>=0.110,<0.116)",
    "uvicorn (>=0.29,<0.34)",
]
flask = [
    "PyYAML (>=6.0)",
    "Flask (>=3.0.2)",
    "gunicorn (>=21.2,<24.0)",
]
all = [
    "PyYAML (>=6.0)",
    "fastapi (>=0.110,<0.116)",
    "uvicorn (>=0.29,<0.34)",
    "Flask (>=3.0.2)",
    "gunicorn (>=21.2,<24.0)",
]

[project.urls]
Homepage = "https://github.com/Chisanan232/PyMock-Server"
Documentation = "https://chisanan232.github.io/PyMock-Server/stable/"
Repository = "https://github.com/Chisanan232/PyMock-Server"
Changelog = "https://chisanan232.github.io/PyMock-Server/latest/release_note/"
"Project Management" = "https://app.clickup.com/9018752317/v/f/90183126979/90182605225"

[project.scripts]
fake = "fake_api_server.runner:run"

# TODO: Let the dependencies for development to be optional
#[tool.poetry.group.dev]
#optional = true

[tool.poetry.group.dev.dependencies]
# Dependency for test
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-html = "^4.1.1"
pytest-rerunfailures = "^14.0"
coverage = "^6.2"
coveralls = "^3.3.1"
# Dependency for CI
pre-commit = "^3.5.0"
pylint = "^3.1.0"
# TODO: Import mypy to check static type of Python project code
#mypy = "^1.2.0"
# Dependency of web framework Flask
Flask = "^3.0.2"
gunicorn = ">=21.2,<24.0"
# Dependency of web framework FastAPI
fastapi = {extras = ["all"], version = ">=0.110,<0.116"}
uvicorn = {extras = ["standard"], version = ">=0.29,<0.34"}
types-PyYAML = "^6.0.12.20241230"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.3"
mkdocs-material = "^9.5.49"
mkdocs-autorefs = "^1.0.1"
mkdocstrings = ">=0.24.1,<0.27.0"
mkdocstrings-python = "^1.9.0"
mike = "^2.1.3"
mkdocs-git-revision-date-localized-plugin = "^1.3.0"
mkdocs-git-committers-plugin-2 = {version = "^2.4.1", python = ">=3.8,<4"}
